+++ libcontainerd/remote/client.go
@@ -25,8 +25,8 @@ import (
 	cerrdefs "github.com/containerd/errdefs"
 	"github.com/containerd/log"
 	"github.com/containerd/typeurl/v2"
+	"github.com/docker/docker/daemon/internal/libcontainerd/queue"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/libcontainerd/queue"
 	libcontainerdtypes "github.com/docker/docker/libcontainerd/types"
 	"github.com/docker/docker/pkg/ioutils"
 	"github.com/hashicorp/go-multierror"
